provided necessary updates, experimented with new functionalities, updated database and inputs files

+++ 19190859_FYP/app/src/main/java/com/example/a19190859_fyp/PerceivedControlInfo.java
@@ -55,15 +55,15 @@ public class PerceivedControlInfo {
         return this.acceleration;
     }
 
-    public double getAccelX(){
+    public float getAccelX(){
         return this.accel_x;
     }
 
-    public double getAccelY(){
+    public float getAccelY(){
         return this.accel_y;
     }
 
-    public double getAccelZ(){
+    public float getAccelZ(){
         return this.accel_z;
     }
